+++ packages/core/empty-state/docs/0-intro.js
@@ -1,14 +1,16 @@
 // @flow
 
 import React from 'react';
-import { md, Example, Props } from '@atlaskit/docs';
+import { md, Example, Props, code } from '@atlaskit/docs';
 
 export default md`
   Empty State is used for various scenarios, for example: empty search, 
   no items, broken link, general error message, welcome screen etc. 
   (usually it takes the whole page).
 
-  ## Examples
+  ## Usage
+
+  ${code`import EmptyState from '@atlaskit/empty-state';`}
 
   ${(
     <Example
@@ -19,5 +21,10 @@ export default md`
     />
   )}
 
-  ${<Props props={require('!!extract-react-types-loader!../src/EmptyState')} />}
+  ${(
+    <Props
+      heading="EmptyState Props"
+      props={require('!!extract-react-types-loader!../src/EmptyState')}
+    />
+  )}
 `;