+++ packages/editor/editor-common/src/schema/nodes/tableNodes.ts
@@ -265,6 +265,7 @@ export const tableCell = {
     '(paragraph | panel | blockquote | orderedList | bulletList | rule | heading | codeBlock |  mediaGroup | mediaSingle | applicationCard | decisionList | taskList | blockCard | extension | unsupportedBlock)+',
   attrs: cellAttrs,
   tableRole: 'cell',
+  marks: 'alignment',
   isolating: true,
   parseDOM: [
     {
@@ -293,6 +294,7 @@ export const tableHeader = {
   attrs: cellAttrs,
   tableRole: 'header_cell',
   isolating: true,
+  marks: 'alignment',
   parseDOM: [
     {
       tag: 'th',