+++ packages/editor/editor-test-helpers/src/index.ts
@@ -82,6 +82,7 @@ export {
   blockCard,
   clean,
   status,
+  alignment,
 } from './schema-builder';
 export { default as defaultSchema } from './schema';
 export * from './html-helpers';
@@ -104,6 +105,6 @@ export {
   bodiedExtensionData,
 } from './mock-extension-data';
 export * from './schema-element-builder';
-export { cardProvider, CardMockProvider } from './card-provider';
+export { cardProvider, EditorTestCardProvider } from './card-provider';
 export * from './enzyme';
 export { compareSelection } from './selection';