+++ packages/core/navigation-next/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/navigation-next
 
+## 3.24.0
+- [minor] [c2c0b0c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c2c0b0c):
+
+  - Adding new component AsyncLayoutManagerWithViewController to be used with code splitting
+
 ## 3.23.0
 - [minor] [b40e33e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b40e33e):